+++ app.js
@@ -3,7 +3,7 @@ import http from 'node:http';
 import express from 'express';
 import cors from 'cors';
 import consola from 'consola';
-import routes from './routes.js';
+import routes from './routes/index.js';
 
 import { socker } from './socker/index.js';
 import { handleError, authenticated } from './middlewares/index.js';
@@ -14,10 +14,7 @@ const server = new http.Server(app);
 socker(server);
 
 app.use(cors({ origin: config.ALLOWLIST_HOSTS, credentials: true }));
-// app.use(cors());
 app.use(express.json());
-// app.use('/users', authenticated);
-// app.use('/search', authenticated);
 
 routes(app);
 
@@ -25,15 +22,7 @@ app.use((error, _request, response, _) => {
   handleError(error, response);
 });
 
-// app.listen(config.ALLOWLIST_HOSTS, () => {
-//   consola.success(`Api listening on port ${config.ALLOWLIST_HOSTS}!`);
-// });
 app.listen(config.API_PORT, () => {
   consola.success(`Api listening on port ${config.API_PORT}!`);
   consola.success(`Swagger docs on http://localhost:${config.API_PORT}/api-docs'`);
 });
-
-// server.listen(config.SOCKET_PORT, () => {
-//   consola.success(`Socker listening on port ${config.SOCKET_PORT}!`);
-//   consola.info(`Api and socker whitelisted for ${config.ALLOWLIST_HOSTS}`);
-// });
